session43/alert/requirements.txt
Flask
Flask-SQLAlchemy
pymysql
python-dotenv
cryptography